requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.5.2
blis==0.7.11
cachetools==5.5.2
catalogue==1.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.11
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.0/en_core_web_sm-2.3.0.tar.gz#sha256=ea8c87848b4a97ced174919e08c00c7888a30495ace38d281d455ee270da2c12
exceptiongroup==1.3.0
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
func_timeout==4.3.5
google-auth==2.40.3
google-auth-oauthlib==0.4.6
grpcio==1.70.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jiter==0.9.1
joblib==1.4.2
kiwisolver==1.4.7
loguru==0.7.3
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
murmurhash==1.0.13
natsort==8.4.0
nltk==3.9.1
numpy==1.24.4
oauthlib==3.3.1
openai==1.93.0
packaging==25.0
pandas==2.0.3
pillow==10.4.0
plac==1.1.3
ply==3.11
preshed==3.0.10
protobuf==3.20.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rapidfuzz==3.9.7
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
spacy==2.3.9
sql_metadata==2.15.0
sqlparse==0.5.3
srsly==1.0.7
thinc==7.4.6
threadpoolctl==3.5.0
tokenizers==0.13.3
tqdm==4.67.1
transformers==4.29.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
wasabi==0.10.1
Werkzeug==3.0.6
zipp==3.20.2
